+++ openpyxl/worksheet/worksheet.py
@@ -47,11 +47,12 @@ from openpyxl.workbook.names.named_range import NamedRange
 from openpyxl.workbook.child import _WorkbookChild
 from openpyxl.utils.bound_dictionary import BoundDictionary
 
+from .datavalidation import DataValidationList
 from .header_footer import HeaderFooter
 from .page import PrintPageSetup, PageMargins, PrintOptions
 from .dimensions import ColumnDimension, RowDimension, DimensionHolder
 from .protection import SheetProtection
-from .filters import AutoFilter
+from .filters import AutoFilter, SortState
 from .views import SheetView, Pane, Selection
 from .properties import WorksheetProperties
 from .pagebreak import PageBreak
@@ -108,10 +109,10 @@ class Worksheet(_WorkbookChild):
         self._images = []
         self._rels = []
         self._drawing = None
-        self._comment_count = 0
+        self._comments = []
         self._merged_cells = []
         self.hyperlinks = set()
-        self._data_validations = []
+        self.data_validations = DataValidationList()
         self.sheet_state = self.SHEETSTATE_VISIBLE
         self.page_setup = PrintPageSetup(worksheet=self)
         self.print_options = PrintOptions()
@@ -121,7 +122,8 @@ class Worksheet(_WorkbookChild):
         self.protection = SheetProtection()
 
         self._current_row = 0
-        self._auto_filter = AutoFilter()
+        self.auto_filter = AutoFilter()
+        self.sort_state = SortState()
         self._freeze_panes = None
         self.paper_size = None
         self.formula_attributes = {}
@@ -189,16 +191,6 @@ class Worksheet(_WorkbookChild):
         """Return an unordered list of the cells in this worksheet."""
         return self._cells.values()
 
-
-    @property
-    def auto_filter(self):
-        """Return :class:`~openpyxl.worksheet.AutoFilter` object.
-
-        `auto_filter` attribute stores/returns string until 1.8. You should change your code like ``ws.auto_filter.ref = "A1:A3"``.
-
-        .. versionchanged:: 1.9
-        """
-        return self._auto_filter
 
     @property
     def freeze_panes(self):
@@ -346,11 +338,6 @@ class Worksheet(_WorkbookChild):
         return self.iter_rows()
 
 
-    @deprecated("Use the max_row property")
-    def get_highest_row(self):
-        return self.max_row
-
-
     @property
     def min_row(self):
         min_row = 1
@@ -371,11 +358,6 @@ class Worksheet(_WorkbookChild):
             rows = set(c[0] for c in self._cells)
             max_row = max(rows)
         return max_row
-
-
-    @deprecated("Use the max_column propery.")
-    def get_highest_column(self):
-        return self.max_column
 
 
     @property
@@ -512,35 +494,6 @@ class Worksheet(_WorkbookChild):
         return tuple(result)
 
 
-    @deprecated("""
-    Use .iter_rows() working with coordinates 'A1:D4',
-    and .get_squared_range() when working with indices (1, 1, 4, 4)
-    and .get_named_range() for named ranges""")
-    def range(self, range_string, row=0, column=0):
-        """Returns a 2D array of cells, with optional row and column offsets.
-
-        :param range_string: cell range string or `named range` name
-        :type range_string: string
-
-        :param row: number of rows to offset
-        :type row: int
-
-        :param column: number of columns to offset
-        :type column: int
-
-        :rtype: tuples of tuples of :class:`openpyxl.cell.Cell`
-
-        """
-        _rs = range_string.upper()
-        m = ABSOLUTE_RE.match(_rs)
-         # R1C1 range
-        if m is not None:
-            rows = self.iter_rows(_rs, row_offset=row, column_offset=column)
-            return tuple(row for row in rows)
-        else:
-            return self.get_named_range(range_string)
-
-
     def set_printer_settings(self, paper_size, orientation):
         """Set printer settings """
 
@@ -555,8 +508,7 @@ class Worksheet(_WorkbookChild):
             object defines the type of data-validation to be applied and the
             cell or range of cells it should apply to.
         """
-        data_validation._sheet = self
-        self._data_validations.append(data_validation)
+        self.data_validations.append(data_validation)
 
     def add_chart(self, chart, anchor=None):
         """
@@ -726,6 +678,8 @@ class Worksheet(_WorkbookChild):
             cols.append(tuple(col))
         return tuple(cols)
 
+
+    @deprecated("Charts and images should be positioned using anchor objects")
     def point_pos(self, left=0, top=0):
         """ tells which cell is under the given coordinates (in pixels)
         counting from the top-left corner of the sheet.